+++ changelog.md
@@ -1,3 +1,20 @@
+# 3.12.0
+
+## Changes
+
+Updated for patch 10.2.7.
+
+### Comms throttle
+
+Voting Frame data is no longer being sent when reloading/reconnecting.
+
+`Add Rolls` option has been reset to `false` due to the extra comms required for it.
+
+## Bugfixes
+
+- *Fixed `/rc council` not properly opening council menu.*
+
+
 # 3.11.1
 
 ## Changes